refactor: modularize Hermes opportunity analyzer
